refactor: rushBalanceData 패칭 및 상태 업데이트하는 로직 훅으로 분리
