+++ client/src/contexts/rushGameContext.tsx
@@ -1,18 +1,13 @@
 import { ReactNode, createContext, useCallback, useEffect, useState } from "react";
-import { useCookies } from "react-cookie";
 import { useLoaderData } from "react-router-dom";
-import { RushAPI } from "@/apis/rushAPI.ts";
 import { CARD_COLOR, CARD_OPTION, CARD_PHASE } from "@/constants/Rush/rushCard";
-import { COOKIE_KEY } from "@/constants/cookie.ts";
-import useFetch from "@/hooks/useFetch.ts";
-import { GetRushBalanceResponse, GetTotalRushEventsResponse } from "@/types/rushApi.ts";
+import { GetTotalRushEventsResponse } from "@/types/rushApi.ts";
 import { CardOption, CardOptionState, GamePhase, RushGameContextType } from "@/types/rushGame";
 import { getMsTime } from "@/utils/getMsTime.ts";
 
 export const RushGameContext = createContext<RushGameContextType | undefined>(undefined);
 
 export const RushGameProvider = ({ children }: { children: ReactNode }) => {
-    const [cookies] = useCookies([COOKIE_KEY.ACCESS_TOKEN]);
     const rushData = useLoaderData() as GetTotalRushEventsResponse;
 
     const [gameState, setGameState] = useState<RushGameContextType["gameState"]>({
@@ -61,32 +56,6 @@ export const RushGameProvider = ({ children }: { children: ReactNode }) => {
         }));
     }, []);
 
-    // TODO: 훅으로 빼기
-    const {
-        data: rushBalanceData,
-        isSuccess: isSuccessRushBalance,
-        fetchData: getRushBalance,
-    } = useFetch<GetRushBalanceResponse, string>((token) => RushAPI.getRushBalance(token));
-
-    const fetchRushBalance = useCallback(async (): Promise<void> => {
-        await getRushBalance(cookies[COOKIE_KEY.ACCESS_TOKEN]);
-    }, [cookies, getRushBalance]);
-
-    useEffect(() => {
-        if (isSuccessRushBalance && rushBalanceData) {
-            const { optionId, leftOption, rightOption } = rushBalanceData;
-
-            setUserSelectedOption(optionId);
-
-            setCardOptions(CARD_OPTION.LEFT_OPTIONS, {
-                selectionCount: leftOption,
-            });
-            setCardOptions(CARD_OPTION.RIGHT_OPTIONS, {
-                selectionCount: rightOption,
-            });
-        }
-    }, [isSuccessRushBalance, rushBalanceData, setUserSelectedOption, setCardOptions]);
-
     // TODO: dispatch action으로 함수로 빼서 RushGame 컴포넌트에서 함수를 호출해서 초기화해주는 것이 맞다고 생각..
     useEffect(() => {
         if (rushData) {
@@ -118,7 +87,6 @@ export const RushGameProvider = ({ children }: { children: ReactNode }) => {
             value={{
                 gameState,
                 setCardOptions,
-                fetchRushBalance,
                 setUserParticipationStatus,
                 setGamePhase,
                 setUserSelectedOption,
